+++ .github/workflows/go.yml
@@ -1,7 +1,6 @@
 name: Go-Build
 on: [push]
 jobs:
-
   build:
     name: Build
     runs-on: ${{ matrix.os }}
@@ -9,15 +8,14 @@ jobs:
       matrix:
         os: [ubuntu-latest, windows-latest, macOS-latest]
     steps:
+      - name: Set up Go 1.12
+        uses: actions/setup-go@v1
+        with:
+          go-version: 1.12
+        id: go
 
-    - name: Set up Go 1.12
-      uses: actions/setup-go@v1
-      with:
-        go-version: 1.12
-      id: go
+      - name: Check out code into the Go module directory
+        uses: actions/checkout@v1
 
-    - name: Check out code into the Go module directory
-      uses: actions/checkout@v1
-
-    - name: Build
-      run: go build .
+      - name: Build
+        run: go build .
