+++ .gitignore
@@ -1,10 +1,10 @@
 .idea/
 .DS_Store
-whatsappAPI
-*.exe
 files/
-test/
-hosting/config.json
+testing/
+example/config.json
 *.gob
+*.csv
+*.jpg
 *.log
-main
+build/
