--- example/config.json
@@ -1,4 +1,0 @@
-{
-  "sentry": "https://afee5b7806244a5fb3b4b6d1263abe70@sentry.io/1533842",
-  "SAP": "9e8d7bbfbaf81e341460ee1d52128e87"
-}
