--- example/go.mod
@@ -1,11 +1,0 @@
-module whatsappWebAPI/example
-
-go 1.14
-
-require (
-	github.com/Baozisoftware/qrcode-terminal-go v0.0.0-20170407111555-c0650d8dff0f
-	github.com/Rhymen/go-whatsapp v0.1.0
-	github.com/getsentry/sentry-go v0.6.0
-	github.com/gin-contrib/cors v1.3.1
-	github.com/gin-gonic/gin v1.6.2
-)
