+++ image.go
@@ -1,0 +1,35 @@
+package main
+
+import (
+	"log"
+	"os"
+
+	"github.com/Rhymen/go-whatsapp"
+)
+
+func image(v SendImage) string {
+	var folder string
+
+	img, err := os.Open(dir + folder + v.Image)
+	if err != nil {
+		log.Printf("Error reading file: %v\n", err)
+		return "Error"
+	}
+
+	msg := whatsapp.ImageMessage{
+		Info: whatsapp.MessageInfo{
+			RemoteJid: "91" + v.Receiver + "@s.whatsapp.net",
+		},
+		Type:    "image/jpeg",
+		Caption: v.Message,
+		Content: img,
+	}
+
+	msgId, err := wac.Send(msg)
+	if err != nil {
+		log.Printf("Error sending message: to %v --> %v\n", v.Receiver, err)
+		return "Error"
+	}
+
+	return "Message Sent -> " + v.Receiver + " : " + msgId
+}
