+++ whatsapp.go
@@ -1,0 +1,77 @@
+package main
+
+import (
+	"bufio"
+	"encoding/gob"
+	"fmt"
+	"os"
+
+	qrcodeTerminal "github.com/Baozisoftware/qrcode-terminal-go"
+	"github.com/Rhymen/go-whatsapp"
+)
+
+func login(wac *whatsapp.Conn) error {
+	fmt.Print("Enter your number: ")
+	scanner := bufio.NewScanner(os.Stdin)
+	scanner.Scan()
+	text := scanner.Text()
+	fmt.Println("Logging in -> " + text)
+	//load saved session
+	session, err := readSession(text)
+	if err == nil {
+		//restore session
+		session, err = wac.RestoreWithSession(session)
+		if err != nil {
+			return fmt.Errorf("restoring failed: %v\n", err)
+		}
+	} else {
+		//no saved session -> regular login
+		qr := make(chan string)
+		go func() {
+			terminal := qrcodeTerminal.New()
+			terminal.Get(<-qr).Print()
+			// obj := qrcodeTerminal.New2(qrcodeTerminal.ConsoleColors.BrightBlue, qrcodeTerminal.ConsoleColors.BrightGreen, qrcodeTerminal.QRCodeRecoveryLevels.Low)
+			// obj.Get(<-qr).Print()
+		}()
+		session, err = wac.Login(qr)
+		if err != nil {
+			return fmt.Errorf("error during login: %v\n", err)
+		}
+	}
+
+	//save session
+	err = writeSession(session, text)
+	if err != nil {
+		return fmt.Errorf("error saving session: %v\n", err)
+	}
+	return nil
+}
+
+func readSession(s string) (whatsapp.Session, error) {
+	session := whatsapp.Session{}
+	file, err := os.Open(dir + s + ".gob")
+	if err != nil {
+		return session, err
+	}
+	defer file.Close()
+	decoder := gob.NewDecoder(file)
+	err = decoder.Decode(&session)
+	if err != nil {
+		return session, err
+	}
+	return session, nil
+}
+
+func writeSession(session whatsapp.Session, s string) error {
+	file, err := os.Create(dir + s + ".gob")
+	if err != nil {
+		return err
+	}
+	defer file.Close()
+	encoder := gob.NewEncoder(file)
+	err = encoder.Encode(session)
+	if err != nil {
+		return err
+	}
+	return nil
+}
